+++ tests/Python/TestComparison.fs
@@ -546,3 +546,46 @@ let ``test Classes must use identity hashing by default`` () =
 // let ``test GetHashCode with lists works`` () =
 //     ([1; 2].GetHashCode(), [1; 2].GetHashCode()) ||> equal
 //     ([2; 1].GetHashCode(), [1; 2].GetHashCode()) ||> notEqual
+
+[<Fact>]
+let ``GetHashCode with primitives works`` () =
+    ((1).GetHashCode(), (1).GetHashCode()) ||> equal
+    ((2).GetHashCode(), (1).GetHashCode()) ||> notEqual
+    ("1".GetHashCode(), "1".GetHashCode()) ||> equal
+    ("2".GetHashCode(), "1".GetHashCode()) ||> notEqual
+
+[<Fact>]
+let ``Equals with primitives works`` () =
+    (true).Equals(true) |> equal true
+    ('1').Equals('1') |> equal true
+    (1y).Equals(1y) |> equal true
+    (1uy).Equals(1uy) |> equal true
+    (1s).Equals(1s) |> equal true
+    (1).Equals(1) |> equal true
+    (1L).Equals(1L) |> equal true
+    (1u).Equals(1u) |> equal true
+    (1us).Equals(1us) |> equal true
+    (1ul).Equals(1ul) |> equal true
+    (1uL).Equals(1uL) |> equal true
+    (1.f).Equals(1.f) |> equal true
+    (1.).Equals(1.) |> equal true
+    (1.m).Equals(1.m) |> equal true
+    ("1").Equals("1") |> equal true
+
+[<Fact>]
+let ``CompareTo with primitives works`` () =
+    (true).CompareTo(true) |> equal 0
+    ('1').CompareTo('1') |> equal 0
+    (1y).CompareTo(1y) |> equal 0
+    (1uy).CompareTo(1uy) |> equal 0
+    (1s).CompareTo(1s) |> equal 0
+    (1).CompareTo(1) |> equal 0
+    (1L).CompareTo(1L) |> equal 0
+    (1u).CompareTo(1u) |> equal 0
+    (1us).CompareTo(1us) |> equal 0
+    (1ul).CompareTo(1ul) |> equal 0
+    (1uL).CompareTo(1uL) |> equal 0
+    (1.f).CompareTo(1.f) |> equal 0
+    (1.).CompareTo(1.) |> equal 0
+    (1.m).CompareTo(1.m) |> equal 0
+    ("1").CompareTo("1") |> equal 0