+++ src/fable-library-rust/src/FSharp.Collections.fs
@@ -7,12 +7,10 @@ module HashIdentity =
         EqualityComparer<'T>.Create(equals, hasher)
 
     let Structural<'T when 'T: equality> : IEqualityComparer<'T> =
-        EqualityComparer<'T>
-            .Create(LanguagePrimitives.GenericEquality, LanguagePrimitives.GenericHash)
+        EqualityComparer<'T>.Create(LanguagePrimitives.GenericEquality, LanguagePrimitives.GenericHash)
 
     let Reference<'T when 'T: not struct> : IEqualityComparer<'T> =
-        EqualityComparer<'T>
-            .Create(LanguagePrimitives.PhysicalEquality, LanguagePrimitives.PhysicalHash)
+        EqualityComparer<'T>.Create(LanguagePrimitives.PhysicalEquality, LanguagePrimitives.PhysicalHash)
 
 module ComparisonIdentity =
     let FromFunction<'T> comparer : IComparer<'T> = Comparer<'T>.Create(comparer)