+++ src/Fable.Transforms/Global/Compiler.fs
@@ -2,7 +2,7 @@ namespace Fable
 
 module Literals =
     [<Literal>]
-    let VERSION = "5.0.0-alpha.7"
+    let VERSION = "5.0.0-alpha.8"
 
     [<Literal>]
     let JS_LIBRARY_VERSION = "1.10.0"