+++ src/Fable.Cli/CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * [All] Add `StringBuiler.Append(c: char, repeatCount: int)` overload (by @roboz0r)
+* [All] Added primitive types equality and comparison (by @ncave)
+* [All] Updated FCS to latest F# 9.0 (by @ncave)
+* [All] Updated Fable-FCS to latest F# 9.0 (by @ncave)
+* [All] Updated metadata to latest .NET 9.0 (by @ncave)
 
 ## 5.0.0-alpha.5 - 2025-01-09